chore: move onSearch to swap prop
